// src/app/shared/models/provider.ts
export class Provider {
  providerId?: number;
  providerName?: string;
  getterId?: number;
  getterName?: string;
  getterPhoneNumber?: string;
  providerTransactionId?: string;
  providerTransactionDate?: Date;
  providerAmt?: number;
  providerStatus?: string;
  providerDate?: Date;
  providerTDate?: string;
  providerDateString?: string;
}